delete UserRepository
